requirements.txt
flask
pymysql
numpy
openai
requests
python-docx
pypdf
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
jsonify